// src/types/pages/floorDashboard.ts
export interface IBaseItem {
  No: number
  FloorItemNo: number
  Name: string
  PositionX: number
  PositionY: number
  Status: number // 1 = color green, or gray
  Alert: number // use worning iocn
}

interface IFloorDoorResult extends IBaseItem {
  LockStat: number
  ContactStat: number
}

interface IFloorNodeResult extends IBaseItem {}

interface IFloorRegionResult extends IBaseItem {}

interface IFloorInputResult extends IBaseItem {}

interface IFloorOutputResult extends IBaseItem {}

interface IFloorElevatorResult extends IBaseItem {}

interface IFloorRelayResult extends IBaseItem {}

interface IFloorCameraResult extends IBaseItem {}

interface IFloorNvrResult extends IBaseItem {}

interface IFloorChannelResult extends IBaseItem {}

interface IFloorGatewayResult extends IBaseItem {}

interface IFloorLocksetResult extends IBaseItem {}

interface IFloorFacegateResult extends IBaseItem {}

interface IFloorSubnodeResult extends IBaseItem {}

interface IFloorReaderResult extends IBaseItem {}

interface IFloorContGateResult extends IBaseItem {}

interface IFloorContLockResult extends IBaseItem {}

interface IFloorIntercomResult extends IBaseItem {}

interface IFloorTriggerResult extends IBaseItem {}

interface IFloorThreatResult extends IBaseItem {}
// ...
export interface IFloorDashboardItems {
  Node: IFloorNodeResult[]
  Door: IFloorDoorResult[]
  Region: IFloorRegionResult[]
  Input: IFloorInputResult[]
  Output: IFloorOutputResult[]
  Elevator: IFloorElevatorResult[]
  Relay: IFloorRelayResult[]
  Camera: IFloorCameraResult[]
  Nvr: IFloorNvrResult[]
  Channel: IFloorChannelResult[]
  Gateway: IFloorGatewayResult[]
  Lockset: IFloorLocksetResult[]
  Facegate: IFloorFacegateResult[]
  Subnode: IFloorSubnodeResult[]
  Reader: IFloorReaderResult[]
  ContGate: IFloorContGateResult[]
  ContLock: IFloorContLockResult[]
  Intercom: IFloorIntercomResult[]
  Trigger: IFloorTriggerResult[]
  Threat: IFloorThreatResult[]
}